Changing logic for date validation and adding test for it
The code would error out before checking the second date format,
now it should be able to catch the Type and Value errors and continue to
check the other date format. If it gets to the end of the for loop
without returning anything, it will raise the OAIErrorBadArgument
exception.